Commit before remove API code
