+++ Digital.Data/Utilities/Exceptions/HandledException.cs
@@ -1,0 +1,11 @@
+﻿namespace Digital.Data.Utilities.Exceptions;
+
+public class HandledException : Exception
+{
+    public int StatusCode { get; set; }
+
+    public HandledException(int statusCode, string message) : base(message)
+    {
+        StatusCode = statusCode;
+    }
+}
