+++ Digital.Data/Utilities/Exceptions/InputValidationException.cs
@@ -1,0 +1,14 @@
+﻿using System.ComponentModel.DataAnnotations;
+
+namespace Digital.Data.Utilities.Exceptions;
+
+public class InputValidationException : Exception
+{
+    public InputValidationException(IEnumerable<ValidationResult> validationResults, string? message = null) : base(
+        message ?? "Input validation errors")
+    {
+        ValidationResults = validationResults;
+    }
+
+    public IEnumerable<ValidationResult> ValidationResults { get; set; }
+}
