+++ Digital.Data/Utilities/Exceptions/InvalidActionException.cs
@@ -1,0 +1,8 @@
+﻿namespace Digital.Data.Utilities.Exceptions;
+
+public class InvalidActionException : HandledException
+{
+    public InvalidActionException(string? message = null) : base(400, message ?? "Action could not be performed")
+    {
+    }
+}
