+++ Digital.Data/Utilities/Exceptions/InvalidIncludeExpressionException.cs
@@ -1,0 +1,16 @@
+﻿namespace Digital.Data.Utilities.Exceptions;
+
+public class InvalidIncludeExpressionException : HandledException
+{
+    public InvalidIncludeExpressionException(Type declarerType, string propName) : base(400,
+        $"{propName} is not a property of {declarerType.Name} or is not an includable property.")
+    {
+    }
+}
+
+public class InvalidIncludeExpressionException<T> : InvalidIncludeExpressionException
+{
+    public InvalidIncludeExpressionException(string propName) : base(typeof(T), propName)
+    {
+    }
+}
