+++ Digital.Data/Utilities/Exceptions/UnauthorizedException.cs
@@ -1,0 +1,8 @@
+﻿namespace Digital.Data.Utilities.Exceptions;
+
+public class UnauthorizedException: HandledException
+{
+    public UnauthorizedException() : base(401, "Unauthorized")
+    {
+    }
+}
